rename exception to exceptions
because this is the preferred way and more consistent.
