+++ src/paspybin/exceptions.py
